Changes to supply cost by changing purchase order item cost:
The api now allows the user to change the supply cost by changing the cost of the item. When an item has a 'unit_cost' key,
the amount is compared to the supply cost for that supplier. If the two values differ, then the supply.cost is change to the new item 'unit_cost'.
A supply log is also created to keep track of price change