Hydrate updates the quantity
When the quantity is changed via a PUT call, the quantity is correctly routed to the correct country quantity.
Then the change in quality is logged to the supply log, showing the difference. This update also includes the
addition of the 'excludes' attribute to the Resource so that any data returned does not include 'quantity_th'
and 'quantity_kh'